+++ packages/token-metadata/src/tokens/network/testnet.ts
@@ -2,10 +2,10 @@ import { TokenMeta } from '../../types'
 import tokens from '../tokens'
 
 export const testnetSymbolToErc20AddressMap = {
-  INJ: '0x85AbEac4F09762e28a49D7dA91260A46766F4F79',
+  INJ: '0xAD1794307245443B3Cb55d88e79EEE4d8a548C03',
   USDT: '0x87aB3B4C8661e07D6372361211B96ed4Dc36B1B5',
   APE: '0x44C21afAaF20c270EBbF5914Cfc3b5022173FEB7',
-  USDC: '0x85AbEac4F09762e28a49D7dA91260A46766F4F79',
+  USDC: '0x07865c6e87b9f70255377e024ace6630c1eaa37f',
   wETH: '0xdB309Bb079EB419C18fe7D568c61cD2FdB65D9aF',
 }
 
@@ -22,11 +22,14 @@ export const tokensBySymbolForTestnet = (
 
   const tokenMeta = tokens[token] as TokenMeta
 
-  if (testnetAddressFromMap) {
-    if (!tokenMeta.erc20) {
-      return result
+  if (!tokenMeta.erc20 && !tokenMeta.cw20 && !tokenMeta.cw20s) {
+    return {
+      ...result,
+      [token.toUpperCase()]: tokenMeta,
     }
+  }
 
+  if (testnetAddressFromMap) {
     return {
       ...result,
       [token.toUpperCase()]: {
@@ -43,6 +46,19 @@ export const tokensBySymbolForTestnet = (
   const cw20TestnetAddressFromMap = testnetSymbolToCw20AddressMap[cw20TokenKey]
 
   if (cw20TestnetAddressFromMap) {
+    if (tokenMeta.cw20) {
+      return {
+        ...result,
+        [token.toUpperCase()]: {
+          ...tokenMeta,
+          cw20: {
+            ...tokenMeta.cw20,
+            address: cw20TestnetAddressFromMap,
+          },
+        },
+      }
+    }
+
     if (tokenMeta.cw20s) {
       const cw20Maps = tokenMeta.cw20s.map((cw20) =>
         cw20.symbol !== cw20TokenKey
@@ -61,9 +77,10 @@ export const tokensBySymbolForTestnet = (
         },
       }
     }
-
-    return result
   }
 
-  return result
+  return {
+    ...result,
+    [token.toUpperCase()]: tokenMeta,
+  }
 }, {}) as Record<string, TokenMeta>