+++ packages/contracts/CHANGELOG.md
@@ -3,6 +3,38 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.0.110](https://github.com/InjectiveLabs/injective-ts/compare/@injectivelabs/contracts@1.0.109...@injectivelabs/contracts@1.0.110) (2023-02-23)
+
+**Note:** Version bump only for package @injectivelabs/contracts
+
+
+
+
+
+## [1.0.109](https://github.com/InjectiveLabs/injective-ts/compare/@injectivelabs/contracts@1.0.108...@injectivelabs/contracts@1.0.109) (2023-02-23)
+
+**Note:** Version bump only for package @injectivelabs/contracts
+
+
+
+
+
+## [1.0.108](https://github.com/InjectiveLabs/injective-ts/compare/@injectivelabs/contracts@1.0.108-alpha.0...@injectivelabs/contracts@1.0.108) (2023-02-20)
+
+**Note:** Version bump only for package @injectivelabs/contracts
+
+
+
+
+
+## [1.0.108-alpha.0](https://github.com/InjectiveLabs/injective-ts/compare/@injectivelabs/contracts@1.0.107...@injectivelabs/contracts@1.0.108-alpha.0) (2023-02-20)
+
+**Note:** Version bump only for package @injectivelabs/contracts
+
+
+
+
+
 ## [1.0.107](https://github.com/InjectiveLabs/injective-ts/compare/@injectivelabs/contracts@1.0.106...@injectivelabs/contracts@1.0.107) (2023-02-15)
 
 **Note:** Version bump only for package @injectivelabs/contracts