+++ packages/networks/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.0.93](https://github.com/InjectiveLabs/injective-ts/compare/@injectivelabs/networks@1.0.92...@injectivelabs/networks@1.0.93) (2023-02-23)
+
+**Note:** Version bump only for package @injectivelabs/networks
+
+
+
+
+
+## [1.0.92](https://github.com/InjectiveLabs/injective-ts/compare/@injectivelabs/networks@1.0.92-alpha.0...@injectivelabs/networks@1.0.92) (2023-02-20)
+
+**Note:** Version bump only for package @injectivelabs/networks
+
+
+
+
+
+## [1.0.92-alpha.0](https://github.com/InjectiveLabs/injective-ts/compare/@injectivelabs/networks@1.0.91...@injectivelabs/networks@1.0.92-alpha.0) (2023-02-20)
+
+**Note:** Version bump only for package @injectivelabs/networks
+
+
+
+
+
 ## [1.0.91](https://github.com/InjectiveLabs/injective-ts/compare/@injectivelabs/networks@1.0.90...@injectivelabs/networks@1.0.91) (2023-02-15)
 
 **Note:** Version bump only for package @injectivelabs/networks